bugfix: nginx netns patch: moved netns_uniq after fd in ngx_connection_s to avoid invalid read.
In function 'ngx_cache_manager_process_cycle', Nginx uses magic number 4 as the
offset of fd in ngx_connection_s. Our patch changes the offset of fd and
breaks the trick, causes an invalid read when `ngx_event_ident` is
called in cache manager.

This issue is caught by Valgrind when cache manager is involved in test.

Signed-off-by: Yichun Zhang (agentzh) <agentzh@gmail.com>